Rename http/sync.js to http/index.js.
